# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ms the trend API",
        "description": "API documentation for the Ms the trend project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/pincodes": {
            "get": {
                "tags": [
                    "Pincodes"
                ],
                "summary": "Get all pincodes (limited to 50,000)",
                "operationId": "39facdcacda19ee20319f89e6aaadb40",
                "responses": {
                    "200": {
                        "description": "List of pincodes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pincode-details": {
            "get": {
                "tags": [
                    "Pincodes"
                ],
                "summary": "Get details of a specific pincode",
                "operationId": "c601ee9b33ac9208c2206acbd6132370",
                "parameters": [
                    {
                        "name": "pincode",
                        "in": "query",
                        "description": "The pincode to search for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "382350"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pincode details found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pincode details"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Pincode parameter is required"
                    },
                    "404": {
                        "description": "Pincode not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/checkPincode": {
            "post": {
                "tags": [
                    "Pincodes"
                ],
                "summary": "Check if a pincode exists",
                "operationId": "3d1f3c69f5d45bb660832f2bab3f8c71",
                "parameters": [
                    {
                        "name": "pincode",
                        "in": "query",
                        "description": "The pincode to check",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "382350"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pincode found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pincode found"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Pincode parameter is required"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product list with catalogs and colors",
                "operationId": "fdf8ea00d86f6800ac851c172e5b25e5",
                "responses": {
                    "200": {
                        "description": "Product List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product List"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "catalog": {
                                                        "type": "object"
                                                    },
                                                    "colors": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "imgPath": {
                                                        "type": "string",
                                                        "example": "http://localhost/images/catalog/"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product detail by catalog ID",
                "operationId": "8ebdcc918059f28f2ce89a025f213432",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Catalog ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product Detail"
                                        },
                                        "imgPath": {
                                            "type": "string",
                                            "example": "http://localhost/images/product/"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Product 1"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "Red"
                                                    },
                                                    "stock": {
                                                        "properties": {
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 50
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}